+++ src/nucname.cpp
@@ -6,8 +6,27 @@
 #ifndef PYNE_IS_AMALGAMATED
 #include "nucname.h"
 #include "state_map.cpp"
+#include <cctype>
+#include <algorithm>
 #endif
 
+/***String used to check for metastable states***/
+std::string pyne::nucname::metastable_states = "mnopqrstuvxyz";
+
+int pyne::nucname::isomer_id(std::string end_char) {
+    
+    int id = -1;
+    size_t pos = pyne::nucname::metastable_states.find(std::tolower(end_char[0]));
+
+    if (pos != std::string::npos) {
+        id = static_cast<int>(pos) + 1;
+    }
+    else if (pyne::contains_substring(pyne::digits, end_char)) {
+        id = 0;
+    }
+
+    return id;
+}
 
 /*** Constructs the LL to zz Dictionary ***/
 pyne::nucname::name_zz_t pyne::nucname::get_name_zz() {
@@ -584,14 +603,15 @@ int pyne::nucname::id(std::string nuc) {
     if (anum < 0)
       throw NotANuclide(nucstr, anum);
 
-    // Figure out if we are meta-stable or not
+    // Figure out if we are metastable or not
     std::string end_char = pyne::last_char(nucstr);
-    if (end_char == "M")
-      newnuc = (10000 * anum) + 1;
-    else if (pyne::contains_substring(pyne::digits, end_char))
-      newnuc = (10000 * anum);
-    else
-      throw NotANuclide(nucstr, newnuc);
+    int isomer_id = pyne::nucname::isomer_id(end_char);
+    if (isomer_id > -1) {
+        newnuc = (10000 * anum) + isomer_id;
+    }
+    else {
+        throw NotANuclide(nucstr, newnuc);
+    }
 
     // Add the Z-number
     elem_name = pyne::remove_characters(nucstr.substr(0, nuclen-1), pyne::digits);
@@ -670,7 +690,7 @@ std::string pyne::nucname::name(int nuc) {
 
   // Add meta-stable flag
   if (0 < ssss)
-    newnuc += "M";
+    newnuc += std::toupper(pyne::nucname::metastable_states[ssss - 1]);
 
   return newnuc;
 }
@@ -834,7 +854,7 @@ std::string pyne::nucname::zzllaaam(int nuc) {
     newnuc += pyne::to_str(aaa);
   // Add meta-stable flag
   if (0 < ssss)
-    newnuc += "m";
+    newnuc += pyne::nucname::metastable_states[ssss - 1];
   return newnuc;
 }
 
@@ -885,12 +905,13 @@ int pyne::nucname::zzllaaam_to_id(std::string nuc) {
 
   // Figure out if we are meta-stable or not
   std::string end_char = pyne::last_char(nucstr);
-  if (end_char == "M")
-    nucid = (10000 * anum) + 1;
-  else if (pyne::contains_substring(pyne::digits, end_char))
-    nucid = (10000 * anum);
-  else
-    throw NotANuclide(nucstr, nucid);
+  int isomer_id = pyne::nucname::isomer_id(end_char);
+  if (isomer_id > -1) {
+      nucid = (10000 * anum) + isomer_id;
+  }
+  else {
+      throw NotANuclide(nucstr, nucid);
+  }
 
   // Add the Z-number
   elem_name = pyne::remove_characters(nucstr.substr(0, nuclen-1), pyne::digits);
@@ -1108,8 +1129,7 @@ std::string pyne::nucname::serpent(int nuc) {
 
   // Add meta-stable flag
   if (0 < ssss)
-    newnuc += "m";
-
+    newnuc += pyne::nucname::metastable_states[ssss - 1];
   return newnuc;
 }
 
@@ -1160,14 +1180,15 @@ int pyne::nucname::serpent_to_id(std::string nuc) {
   }
   int anum = pyne::to_int(anum_str);
 
-  // Figure out if we are meta-stable or not
+  // Figure out if we are metastable or not
   std::string end_char = pyne::last_char(nucstr);
-  if (end_char == "M")
-    nucid = (10000 * anum) + 1;
-  else if (pyne::contains_substring(pyne::digits, end_char))
-    nucid = (10000 * anum);
-  else
-    throw NotANuclide(nucstr, nucid);
+  int isomer_id = pyne::nucname::isomer_id(end_char);
+  if (isomer_id > -1) {
+      nucid = (10000 * anum) + isomer_id;
+  }
+  else {
+      throw NotANuclide(nucstr, nucid);
+  }
 
   // Add the Z-number
   elem_name = pyne::remove_characters(nucstr.substr(0, nuclen-1), pyne::digits);