+++ tests/test_nucname.py
@@ -9,7 +9,6 @@ from pyne.utils import QAWarning
 warnings.simplefilter("ignore", QAWarning)
 from pyne import nucname
 
-
 def test_name_zz():
     assert nucname.name_zz["He"] == 2
     assert nucname.name_zz["U"] == 92
@@ -417,6 +416,7 @@ def test_id():
     assert nucname.id(95942) == 952420004
 
     assert nucname.id("Am-242m") == 952420001
+    assert nucname.id("Ta-182n") == 731820002
 
     assert nucname.id("he") == 20000000
     assert nucname.id("U") == 920000000
@@ -438,8 +438,12 @@ def test_id():
     assert nucname.id("95-Am-242m") == nucname.id("Am-242m")
     assert nucname.id("94-Pu-239") == nucname.id("Pu-239")
     assert nucname.id("95-Am-242") == nucname.id("Am-242")
+    assert nucname.id("72-Hf-179n") == nucname.id("Hf-179n")
 
     pytest.raises(RuntimeError, nucname.id, "0-H-1")
+    pytest.raises(RuntimeError, nucname.id, "Am-242j")
+    pytest.raises(RuntimeError, nucname.id, "") #nuc.empty()
+    pytest.raises(RuntimeError, nucname.id, -100) #nuc < 0
 
 
 def test_name():
@@ -459,6 +463,8 @@ def test_name():
     assert nucname.name(2440961) == "Cm244M"
     assert nucname.name(2390940) == "Pu239"
     assert nucname.name(2420950) == "Am242"
+
+    assert nucname.name(1820732) == "Ta182N"
 
 
 @pytest.mark.parametrize("case, exp", zip(cases,[
@@ -650,6 +656,8 @@ def test_zzllaaam():
     assert nucname.zzllaaam(2390940) == "94-Pu-239"
     assert nucname.zzllaaam(2420951) == "95-Am-242m"
 
+    assert nucname.zzllaaam(1820732) == "73-Ta-182n"
+
 
 def test_zzllaaam_to_id():
     assert nucname.zzllaaam_to_id("94-Pu-239") == nucname.id("Pu-239")
@@ -658,6 +666,10 @@ def test_zzllaaam_to_id():
     assert nucname.zzllaaam_to_id("94-Pu-239") == nucname.id("Pu-239")
     assert nucname.zzllaaam_to_id("95-Am-242") == nucname.id("Am-242")
     assert nucname.zzllaaam_to_id("95-Am-242m") == nucname.id("Am-242m")
+
+    assert nucname.zzllaaam_to_id("73-Ta-182n") == nucname.id("Ta-182n")
+
+    pytest.raises(RuntimeError, nucname.zzllaaam_to_id, "Ta-182b")
 
 
 def test_mcnp():
@@ -806,6 +818,8 @@ def test_serpent():
 
     assert nucname.serpent(2390940) == "Pu-239"
     assert nucname.serpent(2420951) == "Am-242m"
+
+    assert nucname.serpent(1820732) == "Ta-182n"
 
 
 @pytest.mark.parametrize("val, id",set(zip([